Code clean up. Removed unused test classes.

+++ src/main/java/net/usrlib/cms/user/User.java
@@ -38,7 +38,7 @@ public abstract class User {
 		this.phone = phone;
 	}
 
-	public long getUuid() {
+	public int getUuid() {
 		return uuid;
 	}
